Extract TypeEmbedder and ContractEmbedder from ProgramConverter
ProgramConverter mixed type embedding, contract embedding, and
phase orchestration in one 682-line file. Split type embedding
(embedType/embedFunctionPretype/embedClass and their builder
cascades) into TypeEmbedder, and contract embedding
(embedFormverContract/embedKotlinContract) into ContractEmbedder,
following the existing collaborator pattern (TypeResolver,
PropertyResolver, ParameterResolver). ProgramConverter now
delegates through the unchanged ProgramConversionContext interface.

Also regroups StmtConversionContext.kt, which had grown into a
30-line interface plus ~16 unrelated top-level extension functions,
into coherently named files: Declarations, PropertyAccess,
InlineFunctionCalls, BodyConversion, InvariantCollection, and
Linearization (the last for linearizeImpureBody/linearizePureBody,
which are ProgramConversionContext entry points for the
linearization phase, not statement-conversion helpers).
