+++ formver.compiler-plugin/core/src/org/jetbrains/kotlin/formver/core/conversion/PropertyAccess.kt
@@ -1,0 +1,79 @@
+/*
+ * Copyright 2010-2023 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
+ */
+
+package org.jetbrains.kotlin.formver.core.conversion
+
+import org.jetbrains.kotlin.fir.declarations.utils.isFinal
+import org.jetbrains.kotlin.fir.expressions.FirPropertyAccessExpression
+import org.jetbrains.kotlin.fir.references.symbol
+import org.jetbrains.kotlin.fir.symbols.impl.FirIntersectionOverridePropertySymbol
+import org.jetbrains.kotlin.fir.symbols.impl.FirPropertySymbol
+import org.jetbrains.kotlin.fir.symbols.impl.FirValueParameterSymbol
+import org.jetbrains.kotlin.formver.core.embeddings.properties.ClassPropertyAccess
+import org.jetbrains.kotlin.formver.core.embeddings.properties.PropertyAccessEmbedding
+import org.jetbrains.kotlin.formver.core.embeddings.properties.asPropertyAccess
+import org.jetbrains.kotlin.formver.core.isCustom
+import org.jetbrains.kotlin.utils.addToStdlib.ifTrue
+import org.jetbrains.kotlin.utils.filterIsInstanceAnd
+
+val FirIntersectionOverridePropertySymbol.propertyIntersections
+    get() = intersections.filterIsInstanceAnd<FirPropertySymbol> { it.isVal == isVal }
+
+/**
+ * Tries to find final property symbol actually declared in some class instead of
+ * (potentially) fake property symbol.
+ * Note that if some property is found it is fixed since
+ * 1. there can't be two non-abstract properties which don't subsume each other
+ * in the hierarchy (kotlin disallows that) and final properties can't be abstract;
+ * 2. final property can't subsume other final property as that means final property
+ * is overridden.
+ * //TODO: decide if we leave this lookup or consider it unsafe.
+ */
+fun FirPropertySymbol.findFinalParentProperty(): FirPropertySymbol? =
+    if (this !is FirIntersectionOverridePropertySymbol)
+        (isFinal && !isCustom).ifTrue { this }
+    else propertyIntersections.firstNotNullOfOrNull { it.findFinalParentProperty() }
+
+
+/**
+ * This is a key function when looking up properties.
+ * It translates a kotlin `receiver.field` expression to an `ExpEmbedding`.
+ *
+ * Note that in FIR this `field` may be represented as `FirIntersectionOverridePropertySymbol`
+ * which is necessary when the property could hypothetically inherit from multiple sources.
+ * However, we don't register such symbols in the context when traversing the class.
+ * Hence, some advanced logic is needed here.
+ *
+ * First, we try to find an actual backing field somewhere in the parents of the field with a
+ * dfs-like algorithm on `FirIntersectionOverridePropertySymbol`s (it also should be final).
+ *
+ * If final backing field is not found, we lazily create a getter/setter pair for this
+ * `FirIntersectionOverrideProperty`.
+ */
+fun StmtConversionContext.embedPropertyAccess(accessExpression: FirPropertyAccessExpression): PropertyAccessEmbedding =
+    when (val calleeSymbol = accessExpression.calleeReference.symbol) {
+        is FirValueParameterSymbol -> embedParameter(calleeSymbol).asPropertyAccess()
+        is FirPropertySymbol -> {
+            val type = embedType(calleeSymbol.resolvedReturnType)
+            when {
+                accessExpression.dispatchReceiver != null -> {
+                    val property = calleeSymbol.findFinalParentProperty()?.let {
+                        embedProperty(it)
+                    } ?: embedProperty(calleeSymbol)
+                    ClassPropertyAccess(convert(accessExpression.dispatchReceiver!!), property, type)
+                }
+
+                accessExpression.extensionReceiver != null -> {
+                    val property = embedProperty(calleeSymbol)
+                    ClassPropertyAccess(convert(accessExpression.extensionReceiver!!), property, type)
+                }
+
+                else -> embedLocalProperty(calleeSymbol)
+            }
+        }
+
+        else ->
+            error("Property access symbol $calleeSymbol has unsupported type.")
+    }
