+++ formver.compiler-plugin/core/src/org/jetbrains/kotlin/formver/core/conversion/TypeEmbedder.kt
@@ -1,0 +1,129 @@
+/*
+ * Copyright 2010-2023 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
+ */
+
+package org.jetbrains.kotlin.formver.core.conversion
+
+import org.jetbrains.kotlin.descriptors.isInterface
+import org.jetbrains.kotlin.fir.FirSession
+import org.jetbrains.kotlin.fir.resolve.toClassSymbol
+import org.jetbrains.kotlin.fir.symbols.impl.FirConstructorSymbol
+import org.jetbrains.kotlin.fir.symbols.impl.FirFunctionSymbol
+import org.jetbrains.kotlin.fir.symbols.impl.FirRegularClassSymbol
+import org.jetbrains.kotlin.fir.types.*
+import org.jetbrains.kotlin.formver.core.embeddings.types.*
+import org.jetbrains.kotlin.formver.core.isUnique
+import org.jetbrains.kotlin.formver.core.isManual
+import org.jetbrains.kotlin.formver.core.names.embedName
+
+/**
+ * Embeds Kotlin types (and classes) as [TypeEmbedding]s, pairing with the [PretypeBuilder]
+ * hierarchy in `embeddings/types/`.
+ */
+class TypeEmbedder(private val programCtx: ProgramConversionContext) {
+    private val session: FirSession get() = programCtx.session
+    private val typeResolver: TypeResolver get() = programCtx.typeResolver
+
+    fun embedType(type: ConeKotlinType): TypeEmbedding = buildType { embedTypeWithBuilder(type) }
+
+    /**
+     * Returns an embedding of the class type, with details set.
+     */
+    fun embedClass(symbol: FirRegularClassSymbol): ClassTypeEmbedding {
+        val className = symbol.classId.embedName()
+        typeResolver.lookupClassTypeEmbedding(className)?.let { return it }
+
+        val embedding = typeResolver.getEmbeddingOrExecute(className) {
+            val classEmbedding = buildClassPretype {
+                withName(className)
+            }
+
+            typeResolver.register(classEmbedding, symbol.classKind.isInterface)
+            if (symbol.isManual(session)) {
+                typeResolver.markManual(className)
+            }
+
+            symbol.resolvedSuperTypes.forEach {
+                val superTypeName = embedType(it).pretype.name
+                typeResolver.addSubtypeRelation(className, superTypeName)
+            }
+
+            classEmbedding
+        }
+        symbol.propertySymbols.forEach {
+            programCtx.embedProperty(it)
+        }
+        return embedding
+    }
+
+    // Note: keep in mind that this function is necessary to resolve the name of the function!
+    fun embedFunctionPretype(symbol: FirFunctionSymbol<*>): FunctionTypeEmbedding = buildFunctionPretype {
+        embedFunctionPretypeWithBuilder(symbol)
+    }
+
+    private fun TypeBuilder.embedTypeWithBuilder(type: ConeKotlinType): PretypeBuilder = when {
+        type is ConeErrorType -> error("Encountered an erroneous type: $type")
+        type is ConeTypeParameterType -> {
+            isNullable = true; any()
+        }
+
+        type.isString -> {
+            val stringClassSymbol = type.toClassSymbol(session) as FirRegularClassSymbol
+            stringClassSymbol.propertySymbols.forEach {
+                programCtx.embedProperty(it)
+            }
+            string()
+        }
+
+        type.isUnit -> unit()
+        type.isChar -> char()
+        type.isInt -> int()
+        type.isBoolean -> boolean()
+        type.isNothing -> nothing()
+        type.isSomeFunctionType(session) -> function {
+            check(type is ConeClassLikeType) { "Expected a ConeClassLikeType for a function type, got $type" }
+            type.receiverType(session)?.let { withDispatchReceiver { embedTypeWithBuilder(it) } }
+            type.valueParameterTypesWithoutReceivers(session).forEach { param ->
+                withParam { embedTypeWithBuilder(param) }
+            }
+            withReturnType { embedTypeWithBuilder(type.returnType(session)) }
+        }
+
+        type.canBeNull(session) -> {
+            isNullable = true
+            embedTypeWithBuilder(type.withNullability(false, session.typeContext))
+        }
+
+        type.isAny -> any()
+        type is ConeClassLikeType -> {
+            val classLikeSymbol = type.toClassSymbol(session)
+            if (classLikeSymbol is FirRegularClassSymbol) {
+                existing(embedClass(classLikeSymbol))
+            } else {
+                unimplementedTypeEmbedding(type)
+            }
+        }
+
+        else -> unimplementedTypeEmbedding(type)
+    }
+
+    private fun FunctionPretypeBuilder.embedFunctionPretypeWithBuilder(symbol: FirFunctionSymbol<*>) {
+        symbol.receiverType?.let {
+            withDispatchReceiver { embedTypeWithBuilder(it) }
+        }
+        symbol.extensionReceiverType?.let {
+            withExtensionReceiver { embedTypeWithBuilder(it) }
+        }
+        symbol.valueParameterSymbols.forEach { param ->
+            withParam {
+                embedTypeWithBuilder(param.resolvedReturnType)
+            }
+        }
+        withReturnType { embedTypeWithBuilder(symbol.resolvedReturnType) }
+        returnsUnique = symbol.isUnique(session) || symbol is FirConstructorSymbol
+    }
+
+    private fun TypeBuilder.unimplementedTypeEmbedding(type: ConeKotlinType): PretypeBuilder =
+        programCtx.handleUnsupportedFeature(null, "The embedding for type $type is not yet implemented.") { unit() }
+}
